menu-minglers/pyproject.toml
[tool.poetry]
name = "menu-minglers"
version = "0.1.0"
description = "A FastAPI application for BärnHäckt MenuMingle"
authors = ["Christoph Keller <christoph.keller@bambit.ch>"]
readme = "README.md"
packages = [{include = "app"}]

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.115.14"
uvicorn = {extras = ["standard"], version = "^0.32.1"}
pydantic = "^2.10.4"
pydantic-settings = "^2.8.0"
python-dotenv = "^1.0.0"
websockets = "^13.0"
tinytroupe = { git = "https://github.com/microsoft/TinyTroupe.git", tag = "v0.5.2" }
torch = { version = "2.7.1+cpu", source = "pytorch-cpu" }

[tool.poetry.group.dev.dependencies]
pytest = "^8.0.0"
pytest-asyncio = "^0.23.0"
httpx = "^0.27.0"
black = "^24.1.0"
isort = "^5.13.0"
flake8 = "^7.0.0"
mypy = "^1.8.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ['py312']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[[tool.poetry.source]]
name = "pytorch-cpu"
url = "https://download.pytorch.org/whl/cpu"
priority = "explicit"

[tool.isort]
profile = "black"
multi_line_output = 3
line_length = 88

[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
